fixing some warnings
git-svn-id: https://openlierox.svn.sourceforge.net/svnroot/openlierox@4216 34602234-ff1f-0410-a465-ea8f3b77ab7f

+++ libs/hawknl/include/nl.h
@@ -139,7 +139,7 @@ typedef struct nl_cond_t *NLcond;
 /* misc. */
 typedef int NLint;
 typedef unsigned int NLuint;
-typedef unsigned int NLenum;
+typedef unsigned long NLenum;
 typedef void NLvoid;
 typedef NLlong NLsocket;
 /* NOTE: NLchar is only to be used for external strings
